Moving MongoDBSchemaStore into correct namespace an folder
